ai_service/requirements.txt
# מחזיק את כל הספריות שהפרויקט צריך בשביל לרוץ

fastapi
uvicorn
openai
python-dotenv
pydantic
requests
